+++ data-raw/packages.R
@@ -16,6 +16,7 @@ paletteer_packages <- tibble::tribble(
   "futurevisions",      "JoeyStanley/futurevisions",        "0.1.1",       FALSE, NA,
   "gameofthrones",      "aljrico/gameofthrones",            "1.0.2",       TRUE,  "1.0.2",
   "ggpomological",      "gadenbuie/ggpomological",          "0.1.2",       FALSE, NA,
+  "ggprism",            "csdaw/ggprism",                    "1.0.3.9000",  TRUE,  "1.0.3",
   "ggsci",              "nanxstats/ggsci",                  "2.9",         TRUE,  "2.9",
   "ggthemes",           "jrnold/ggthemes",                  "4.2.0",       TRUE,  "4.2.0",
   "ggthemr",            "Mikata-Project/ggthemr",           "1.1.0",       FALSE, NA,